+++ modules/os2forms_digital_post/src/Helper/DigitalPostHelper.php
@@ -76,7 +76,8 @@ final class DigitalPostHelper implements LoggerInterface {
 
     // RecipientID should be the same in Message and Forsendelse,
     // so fetch it from Message as it is always set.
-    $msg = sprintf('Sent digital post of type %s to %s', $type, $message->getMessageHeader()->getRecipient()->getRecipientID());
+    $msg = sprintf('Sent digital post of type %s to %s.', $type, $message->getMessageHeader()->getRecipient()->getRecipientID());
+    $msg .= $submission ? sprintf(' Webform id %s.', $submission->getWebform()->id()) : '';
     $this->auditLogger->info('DigitalPost', $msg);
 
     return [$response, $service->getLastKombiMeMoMessage()];