+++ CHANGELOG.md
@@ -11,6 +11,8 @@ before starting to add changes. Use example [placed in the end of the page](#exa
 
 ## [Unreleased]
 
+- Added webform ID to digital post audit logging messages.
+
 ## [3.19.0] 2024-12-06
 
 - Ensured installation of `os2web_audit`