+++ CHANGELOG.md
@@ -11,6 +11,12 @@ before starting to add changes. Use example [placed in the end of the page](#exa
 
 ## [Unreleased]
 
+## [3.20.0] 2024-12-09
+
+- Added webform ID to digital post audit logging messages.
+- Added audit logging to `os2forms_fasit`
+- Added audit logging to `os2forms_fbs_handler`
+
 ## [3.19.0] 2024-12-06
 
 - Ensured installation of `os2web_audit`
@@ -308,7 +314,8 @@ before starting to add changes. Use example [placed in the end of the page](#exa
 - Security in case of vulnerabilities.
 ```
 
-[Unreleased]: https://github.com/OS2Forms/os2forms/compare/3.19.0...HEAD
+[Unreleased]: https://github.com/OS2Forms/os2forms/compare/3.20.0...HEAD
+[3.20.0]: https://github.com/OS2Forms/os2forms/compare/3.19.0...3.20.0
 [3.19.0]: https://github.com/OS2Forms/os2forms/compare/3.18.0...3.19.0
 [3.18.0]: https://github.com/OS2Forms/os2forms/compare/3.17.0...3.18.0
 [3.17.0]: https://github.com/OS2Forms/os2forms/compare/3.16.2...3.17.0